+++ packages/nextjs/jest.config.js
@@ -4,5 +4,5 @@ module.exports = {
   ...baseConfig,
   // This prevents the build tests from running when unit tests run. (If they do, they fail, because the build being
   // tested hasn't necessarily run yet.)
-  testPathIgnorePatterns: ['<rootDir>/test/buildProcess/'],
+  testPathIgnorePatterns: ['<rootDir>/test/buildProcess/', '<rootDir>/test/integration/'],
 };
