+++ packages/nextjs/test/integration/test/server/excludedApiEndpoints.test.ts
@@ -1,4 +1,4 @@
-import { NextTestEnv } from '../utils/server';
+import { NextTestEnv } from './utils/helpers';
 
 describe('Excluded API Endpoints', () => {
   it('Should exclude API endpoint via RegExp', async () => {
