+++ packages/nextjs/test/integration/test/server/tracing200.test.ts
@@ -1,4 +1,4 @@
-import { NextTestEnv } from '../utils/server';
+import { NextTestEnv } from './utils/helpers';
 
 describe('Tracing 200', () => {
   it('should capture a transaction', async () => {
