+++ packages/nextjs/test/integration/test/server/tracingServerGetInitialProps.test.ts
@@ -1,4 +1,4 @@
-import { NextTestEnv } from '../utils/server';
+import { NextTestEnv } from './utils/helpers';
 
 describe('getInitialProps', () => {
   it('should capture a transaction', async () => {
