+++ packages/nextjs/test/integration/test/server/tracingServerGetServerSideProps.test.ts
@@ -1,4 +1,4 @@
-import { NextTestEnv } from '../utils/server';
+import { NextTestEnv } from './utils/helpers';
 
 describe('getServerSideProps', () => {
   it('should capture a transaction', async () => {
