+++ packages/nextjs/tsconfig.json
@@ -4,7 +4,6 @@
   "include": ["src/**/*"],
 
   "compilerOptions": {
-    "esModuleInterop": true
     // package-specific options
   }
 }
