+++ packages/nextjs/tsconfig.test.json
@@ -4,6 +4,7 @@
   "include": ["test/**/*"],
 
   "compilerOptions": {
+    "esModuleInterop": true,
     // should include all types from `./tsconfig.json` plus types for all test frameworks used
     "types": ["node", "jest"]
